Download and crop a netcdf tile from the 1-km DAYMET daily weather dataset.

get_daymet_tile returns a list of RasterBricks---one for each element---cropped within the specified template. If template is not provided, returns the entire DAYMET tile.

get_daymet_tile(template, tileID, elements = NULL, years = NULL, raw.dir)

Arguments

template
A Raster* or Spatial* object to serve as a template for cropping. If missing, entire tile is returned.

tileID
A numeric indicating the DAYMET tile ID number.

elements
A character vector of elements to extract. The available elements are: dayl = Duration of the daylight period in seconds per day. This calculation is based on the period of the day during which the sun is above a hypothetical flat horizon. prcp = Daily total precipitation in millimeters per day, sum of all forms converted to water-equivalent. Precipitation occurrence on any given day may be ascertained. srad = Incident shortwave radiation flux density in watts per square meter, taken as an average over the daylight period of the day. NOTE: Daily total radiation (MJ/m2/day) can be calculated as follows: ((srad (W/m2) * dayl (s/day)) / l,000,000) swe = Snow water equivalent in kilograms per square meter. The amount of water contained within the snowpack. tmax = Daily maximum 2-meter air temperature in degrees Celsius. tmin = Daily minimum 2-meter air temperature in degrees Celsius. vp = Water vapor pressure in pascals. Daily average partial pressure of water vapor.

years
A numeric vector of years to extract.

raw.dir
A character string indicating where raw downloaded files should be put. The directory will be created if missing. Defaults to './RAW/NED/'.

Value

A a list of RasterBricks---one for each element---cropped within the specified template.