Download and crop the National Land Cover Database.

get_nlcd returns a RasterLayer of NLCD data cropped to a given template study area.

get_nlcd_new(template, label, year = 2016, dataset = "landcover",
  raw.dir = "./RAW/NLCD", extraction.dir = paste0("./EXTRACTIONS/",
  label, "/NLCD"), raster.options = c("COMPRESS=DEFLATE", "ZLEVEL=9",
  "INTERLEAVE=BAND"), force.redo = F)

Arguments

template
A Raster* or Spatial* object to serve as a template for cropping.

label
A character string naming the study area.

year
An integer representing the year of desired NLCD product. Acceptable values are 2016 (default), 2013, 2011, 2006, and 2001.

dataset
A character string representing type of the NLCD product. Acceptable values are 'landcover' (default), 'impervious', and 'canopy'. As of February 7, 2018, the canopy data for 2006 are not available through the National Map Staged datasets, and so aren't available in FedData.

raw.dir
A character string indicating where raw downloaded files should be put. The directory will be created if missing. Defaults to './RAW/NLCD/'.

extraction.dir
A character string indicating where the extracted and cropped DEM should be put. The directory will be created if missing. Defaults to './EXTRACTIONS/NLCD/'.

raster.options
a vector of options for raster::writeRaster.

force.redo
If an extraction for this template and label already exists, should a new one be created?

Value

A RasterLayer DEM cropped to the extent of the template.

Examples

# NOT RUN {
# Extract data for the Village Ecodynamics Project 'VEPIIN' study area:
# http://village.anth.wsu.edu
vepPolygon <- polygon_from_extent(raster::extent(672800,740000,4102000,4170000),
     proj4string='+proj=utm +datum=NAD83 +zone=12')

# Get the NLCD (USA ONLY)
# Returns a raster
NLCD <- get_nlcd(template=vepPolygon, label='VEPIIN')

# Plot with raster::plot
plot(NLCD)
# }