Download and crop the 1-km DAYMET daily weather dataset.

get_daymet returns a RasterBrick of weather data cropped to a given template study area.

get_daymet(
  template,
  label,
  elements = c("dayl", "prcp", "srad", "swe", "tmax", "tmin", "vp"),
  years = 1980:(lubridate::year(Sys.time()) - 1),
  region = "na",
  tempo = "day",
  extraction.dir = paste0(tempdir(), "/FedData/extractions/daymet/", label, "/"),
  raster.options = c("COMPRESS=DEFLATE", "ZLEVEL=9", "INTERLEAVE=BAND"),
  force.redo = F,
  progress = TRUE
)

Arguments

template
A Raster* or Spatial* object to serve as a template for cropping.

label
A character string naming the study area.

elements
A character vector of elements to extract.
The available elements are:
dayl = Duration of the daylight period in seconds per day. This calculation is based on the period of the day during which the sun is above a hypothetical flat horizon.
prcp = Daily total precipitation in millimeters per day, sum of all forms converted to water-equivalent. Precipitation occurrence on any given day may be ascertained.
srad = Incident shortwave radiation flux density in watts per square meter, taken as an average over the daylight period of the day. NOTE: Daily total radiation (MJ/m2/day) can be calculated as follows: ((srad (W/m2) * dayl (s/day)) / l,000,000)
swe = Snow water equivalent in kilograms per square meter. The amount of water contained within the snowpack.
tmax = Daily maximum 2-meter air temperature in degrees Celsius.
tmin = Daily minimum 2-meter air temperature in degrees Celsius.
vp = Water vapor pressure in pascals. Daily average partial pressure of water vapor.

years
A numeric vector of years to extract.

region
The name of a region. The available regions are:
na = North America
hi = Hawaii
pr = Puerto Rico

tempo
The frequency of the data. The available tempos are:
day = Daily data
mon = Monthly summary data
ann = Annual summary data

extraction.dir
A character string indicating where the extracted and cropped DEM should be put. Defaults to a temporary directory.

raster.options
a vector of options for raster::writeRaster.

force.redo
If an extraction for this template and label already exists in extraction.dir, should a new one be created?

progress
Draw a progress bar when downloading?

Value

A named list of RasterBricks of weather data cropped to the extent of the template.

Examples

if (FALSE) {

# Get the DAYMET (North America only)
# Returns a list of raster bricks
DAYMET <- get_daymet(
  template = FedData::meve,
  label = "meve",
  elements = c("prcp", "tmin", "tmax"),
  years = 1980:1985
)

# Plot with raster::plot
plot(DAYMET$tmin$X1985.10.23)
}