Download the latest version of the ITRDB, and extract given parameters.

get_itrdb returns a named list of length 3:

'metadata': A data.table or SpatialPointsDataFrame (if makeSpatial==TRUE) of the locations and names of extracted ITRDB chrononlogies,

'widths': A matrix of tree-ring widths/densities given user selection, and

'depths': A matrix of tree-ring sample depths.

get_itrdb(template = NULL, label = NULL, recon.years = NULL,
  calib.years = NULL, species = NULL, measurement.type = NULL,
  chronology.type = NULL, makeSpatial = F, raw.dir = "./RAW/ITRDB",
  extraction.dir = ifelse(!is.null(label), paste0("./EXTRACTIONS/", label,
  "/ITRDB"), "./EXTRACTIONS/ITRDB"), force.redo = FALSE)

Arguments

template
A Raster* or Spatial* object to serve as a template for selecting chronologies. If missing, all available global chronologies are returned.

label
A character string naming the study area.

recon.years
A numeric vector of years over which reconstructions are needed; if missing, the union of all years in the available chronologies are given.

calib.years
A numeric vector of all required years---chronologies without these years will be discarded; if missing, all available chronologies are given.

species
A character vector of 4-letter tree species identifiers; if missing, all available chronologies are given.

measurement.type
A character vector of measurement type identifiers. Options include:

'Total Ring Density'

'Earlywood Width'

'Earlywood Density'

'Latewood Width'

'Minimum Density'

'Ring Width'

'Latewood Density'

'Maximum Density'

'Latewood Percent'

if missing, all available chronologies are given.

chronology.type
A character vector of chronology type identifiers. Options include:

'ARSTND'

'Low Pass Filter'

'Residual'

'Standard'

'Re-Whitened Residual'

'Measurements Only'

if missing, all available chronologies are given.

makeSpatial
Should the metadata be presented as a SpatialPointsDataFrame? Defaults to FALSE.

raw.dir
A character string indicating where raw downloaded files should be put. The directory will be created if missing. Defaults to './RAW/ITRDB/'.

extraction.dir
A character string indicating where the extracted and cropped ITRDB dataset should be put. The directory will be created if missing. Defaults to './EXTRACTIONS/ITRDB/'.

force.redo
If an extraction already exists, should a new one be created? Defaults to FALSE.

Value

A named list containing the 'metadata', 'widths', and 'depths' data.

Examples

# NOT RUN {
# Extract data for the Village Ecodynamics Project 'VEPIIN' study area:
# http://village.anth.wsu.edu
vepPolygon <- polygon_from_extent(raster::extent(672800,740000,4102000,4170000),
     proj4string='+proj=utm +datum=NAD83 +zone=12')

# Get the ITRDB records
ITRDB <- get_itrdb(template=vepPolygon, label='VEPIIN', makeSpatial=T)

# Plot the VEP polygon
plot(vepPolygon)

# Map the locations of the tree ring chronologies
plot(ITRDB$metadata, pch=1, add=T)
legend('bottomleft', pch=1, legend='ITRDB chronologies')
# }