Download and crop the PAD-US Dataset.

get_padus returns a list of sf objects extracted from the PAD-US Dataset. Data are retrieved directly from PAD-US ArcGIS Web Services.

Usage

get_padus(
  template,
  label,
  layer = c("Manager_Name"),
  extraction.dir = file.path(tempdir(), "FedData", "extractions", "padus", label),
  force.redo = FALSE
)

Arguments

template

An Simple Feature or SpatRaster object to serve as a template for cropping. Optionally, a vector of unit names, e.g., c('Mesa Verde National Park','Ute Mountain Reservation') may be provided.

label

A character string naming the study area.

layer

A character vector containing one or more PAD-US Layers. By default, the Manager_Name layer is downloaded.

Protection_Status_by_GAP_Status_Code: PAD-US 3.0 Protection Status by GAP Status Code — Service representing a measure of management intent to permanently protect biodiversity. GAP 1&2 areas are primarily managed for biodiversity, GAP 3 are managed for multiple uses including conservation and extraction, GAP 4 no known mandate for biodiversity protection. GAP Status Codes 1-3 are displayed, GAP 4 areas included but not displayed.

Public_Access: PAD-US 3.0 Public Access — Service representing general level of public access permitted in the area - Open, Restricted (permit, seasonal), Closed. Public Access Unknown areas not displayed. Use to show general categories of public access (however, not all areas have been locally reviewed).

Fee_Manager: PAD-US 3.0 Fee Manager — Manager or administrative agency names standardized nationally. Use for categorization by manager name, with detailed federal managers and generic state/local/other managers. Where available this layer includes fee simple parcels from the Fee feature class plus DOD and Tribal areas from the Proclamation feature class.

Manager_Name: PAD-US 3.0 Manager Name — Service representing coarse level land manager description from "Agency Type" Domain, "Manager Type" Field (for example, Federal, Tribal, State, Local Gov, Private). Use for broad categorization of manager levels, for general depictions of who manages what areas.

Manager_Type: PAD-US 3.0 Manager Type — Service representing coarse level land manager description from "Agency Type" Domain, "Manager Type" Field (for example, Federal, Tribal, State, Local Gov, Private). Use for broad categorization of manager levels, for general depictions of who manages what areas.

Federal_Fee_Managers_Authoritative: PAD-US 3.0 Federal Fee Managers Authoritative — An ArcGIS WebService describing authoritative fee data for federal managers or administrative agencies by name. U.S. Department of Defense and Tribal areas shown from the Proclamation feature class. Use to depict authoritative fee data for individual federal management agencies (no state, local or private lands). This service does not include designations that often overlap state, private or other inholdings. U.S. Department of Defense internal land ownership is not represented but is implied Federal. See the Federal Management Agencies service for a combined view of fee ownership, designations, and easements.

Federal_Management_Agencies: PAD-US 3.0 Federal Management Agencies — Federal managers or administrative agencies by name. Use to depict individual federal management agencies (no state, local or private lands). This map is based on the Combined Proclamation, Marine, Fee, Designation, Easement feature class.

Protection_Mechanism_Category: PAD-US 3.0 Protection Mechanism Category — Service representing the protection mechanism category including fee simple, internal management designations, easements, leases and agreements, and Marine Areas. Use to show categories of land tenure for all protected areas, including marine areas.

Proclamation_and_Other_Planning_Boundaries: PAD-US 3.0 Proclamation and Other Planning Boundaries — Service representing boundaries that provide additional context. Administrative agency name standardized for the nation (DOD, FWS, NPS, USFS, Tribal). Boundaries shown with outline only, as proclamation data do not depict actual ownership or management. Use to show outline of agency proclamation, approved acquisition or other planning boundaries where internal ownership is not depicted.

Fee_Topology_Overlaps: PAD-US 3.0 Topology Overlaps — Topology assessment of the Fee feature class. Use to identify overlaps in Fee data between Federal agencies and between Federal/State lands.

extraction.dir

A character string indicating where the extracted and cropped PAD-US data should be put.

force.redo

If an extraction for this template and label already exists, should a new one be created?

Value

A list of sf::sf collections extracted from the PAD-US Dataset.

Details

PAD-US is America’s official national inventory of U.S. terrestrial and marine protected areas that are dedicated to the preservation of biological diversity and to other natural, recreation and cultural uses, managed for these purposes through legal or other effective means. PAD-US also includes the best available aggregation of federal land and marine areas provided directly by managing agencies, coordinated through the Federal Geographic Data Committee Federal Lands Working Group.

Examples

if (FALSE) { # \dontrun{
# Get the PAD-US (USA ONLY)
PADUS <- get_padus(
  template = FedData::meve,
  label = "meve"
)
PADUS
} # }